+++ docs/changelog.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Fixed `six-checkbox` checkbox-input being distorted with multiline content/text
 - Fixed `six-sidebar-item-group` to wrap the sidebar text if it gets too long
+- Updated the `six-language-switcher` component to include tabindex attributes for better
+  accessibility.
 
 ## 4.2.1 - 2024-03-14